requirements.txt
replicate>=0.22.0
python-dotenv>=1.0.0
requests>=2.31.0
Flask>=3.0.0
Flask-Cors>=4.0.0
